+++ sample/sample.js
@@ -1,7 +1,7 @@
 
-let Hjson=require("hjson");
-let fs=require("fs");
-let text=fs.readFileSync("test.hjson", "utf8");
+let Hjson=require('hjson');
+let fs=require('fs');
+let text=fs.readFileSync('test.hjson', 'utf8');
 
 // parse either JSON or Hjson
 let data=Hjson.parse(text);
@@ -9,19 +9,19 @@ console.log(data.hello);
 console.log();
 
 // convert to JSON
-console.log("--- JSON output:");
+console.log('--- JSON output:');
 console.log(JSON.stringify(data, null, 2));
 console.log();
 
 // convert to Hjson
-console.log("\n--- Hjson output:");
+console.log('\n--- Hjson output:');
 console.log(Hjson.stringify(data));
 
 // parse, keep whitespace and comments
 data=Hjson.rt.parse(text);
 
 // modify like you normally would
-data.foo="text";
+data.foo='text';
 
-console.log("\n--- Hjson output with comments:");
+console.log('\n--- Hjson output with comments:');
 console.log(Hjson.rt.stringify(data));